Initial project creation
The contents of this project are based on the files in the OpenPOWER
Foundation Docs-Template project at:
https://github.com/OpenPOWERFoundation/Docs-Template

Specifically, the files in the common directory of this project were
taken explicitly from the docs/common directory and the pom.xml file
from the docs/pom.xml (with a minor tweak).  Licensing is preserved.
A new README.md file was created.

As mentioned the the README.md file, this project is the basis for
all OpenPOWER documentation projects and includes the master POM
and common source files.  Thus, the components are all extracted from
existing places.

Signed-off-by: Jeff Scheel <scheel@us.ibm.com>